+++ arch/sim/sim.cc
@@ -1173,7 +1173,7 @@ uint32_t arch_trace_now(void)
 }
 
 // No host core clock to report: 0 == unknown (the ABI contract).
-uint32_t arch_cpu_clock_hz(void)
+uint64_t arch_cpu_clock_hz(void)
 {
     return 0;
 }
